Rename Booster Ramp HLA package and make script work. Add minor fixes.
