+++ pyqt-apps/scripts/sirius-hla-bo-ap-ramp.py
@@ -4,14 +4,15 @@
 
 import sys
 import argparse as _argparse
+from siriuspy.envars import vaca_prefix as _vaca_prefix
 from siriushla.sirius_application import SiriusApplication
-from siriushla.bo_ap_ramp.ramp_main import RampMain
+from siriushla.bo_ap_ramp.main_window import RampMain
 from siriushla import util
 
 parser = _argparse.ArgumentParser(
     description="Run Booster Ramp HLA Interface.")
 parser.add_argument('-p', "--prefix", type=str,
-                    default='ca://fernando-lnls452-linux-AS-Glob:TI-EVG:',
+                    default=_vaca_prefix,
                     help="Define the prefix for the PVs in the window.")
 args = parser.parse_args()
