+++ pyqt-apps/setup.py
@@ -25,11 +25,12 @@ setup(
     scripts=[
         'scripts/sirius-hla-as-ap-config-manager.py',
         'scripts/sirius-hla-as-ma-launcher.py',
+        'scripts/sirius-hla-as-ps-cycle.py',
         'scripts/sirius-hla-as-ps-launcher.py',
         'scripts/sirius-hla-as-ti-control.py',
-        'scripts/sirius-hla-as-ps-cycle.py',
         'scripts/sirius-hla-bo-ap-currlt.py',
         'scripts/sirius-hla-bo-ap-chromcorr.py',
+        'scripts/sirius-hla-bo-ap-ramp.py',
         'scripts/sirius-hla-bo-ap-tunecorr.py',
         'scripts/sirius-hla-bo-config-manager.py',
         'scripts/sirius-hla-bo-ma-control.py',
@@ -45,7 +46,6 @@ setup(
         'scripts/sirius-hla-ts-ap-control.py',
         'scripts/sirius-hla-ts-ap-posang.py',
         'scripts/sirius-hla-ts-ma-control.py',
-        'scripts/sirius-hla-bo-ramp.py',
         ],
     zip_safe=False
     )
