+++ pyqt-apps/siriushla/bo_ap_ramp/__init__.py
@@ -1,0 +1,3 @@
+from siriushla.bo_ap_ramp import *
+
+__all__ = ('main_window', )
