+++ pyqt-apps/siriushla/bo_ap_ramp/auxiliar_classes.py
