--- pyqt-apps/siriushla/bo_ramp/__init__.py
@@ -1,4 +1,0 @@
-from siriushla.bo_ramp import *
-
-__all__ = ('ramp_main, ramp_settings, ramp_commands, ramp_params',
-           'optics_adjust, ramp_statistics')
